+++ utils/retry/retryer_test.go
@@ -18,6 +18,8 @@ func TestNewRetryer(t *testing.T) {
 		MaxRetries: 3,
 		MinDelay:   50 * time.Millisecond,
 		MaxDelay:   2*time.Second + 500*time.Millisecond,
+		MinJitter:  50 * time.Millisecond,
+		MaxJitter:  250 * time.Millisecond,
 	}
 
 	require.Equal(t, Retryer[int]{options: options}, retryer)